Add LCD library and implement SPI communication for MEGA and UNO
- Added lcd.h library for HD44780U LCD display support.
- Implemented main.c for MEGA with keypad and LCD initialization, SPI master communication.
- Created stdutils.h for standard utility macros and data types.
- Developed main.c for UNO as SPI slave with USART communication.
- Updated TODO.md with project progress and future tasks.
