+++ Project_MEGA/Debug/Project_MEGA.hex
@@ -1,0 +1,108 @@
+:10000000E9C00000FEC00000FCC00000FAC0000013
+:10001000F8C00000F6C00000F4C00000F2C000000C
+:10002000F0C00000EEC00000ECC00000EAC000001C
+:10003000E8C00000E6C00000E4C00000E2C000002C
+:10004000E0C00000DEC00000DCC00000DAC000003C
+:10005000D8C00000D6C00000D4C00000D2C000004C
+:10006000D0C00000CEC00000CCC00000CAC000005C
+:10007000C8C00000C6C00000C4C00000C2C000006C
+:10008000C0C00000BEC00000BCC00000BAC000007C
+:10009000B8C00000B6C00000B4C00000B2C000008C
+:1000A000B0C00000AEC00000ACC00000AAC000009C
+:1000B000A8C00000A6C00000A4C00000A2C00000AC
+:1000C000A0C000009EC000009CC000009AC00000BC
+:1000D00098C0000096C0000094C0000092C00000CC
+:1000E00090C0000077017F017F017F0179017F01CE
+:1000F0007B017D017F017F017F017F017F017F0106
+:100100007F017F017F017F017F017F017F017F01EF
+:100110007F017F017F017F017F017F017F017F01DF
+:100120007F017F017F017F017F017F017F017F01CF
+:100130007F017F017F017F017F017F017F017F01BF
+:100140007F017F017F017F017F017F017F017F01AF
+:100150007F017F017F017F017F017F017F017F019F
+:100160007F017F016F017F017F017F0171017F01AD
+:10017000730175017F017F017F017F017F017F0195
+:100180007F017F017F017F017F017F017F017F016F
+:100190007F017F017F017F017F017F017F017F015F
+:1001A0007F017F0167017F017F017F0169017F017D
+:1001B0006B016D017F017F017F017F017F017F0165
+:1001C0007F017F0161017F017F017F0181017F014B
+:1001D0006301650111241FBECFEFD1E2DEBFCDBFA9
+:1001E00000E00CBF12E0A0E0B2E0EAE8F6E000E0D8
+:1001F0000BBF02C007900D92A431B107D9F7DCD133
+:1002000042C2FECE20E09FEE10C090930801E9EFBD
+:10021000F0E03197F1F700C00000809106018F7087
+:100220008F3029F4990F9F5F2F5F243070F3907FF8
+:10023000890F089580EF8093070108958FE08093E0
+:100240000801809106018F708F30C1F789EF90E02F
+:100250000197F1F700C000008FE0809308018091C2
+:1002600006018F708F3051F708958FE08093080159
+:10027000809106018F708F30C1F389EF90E0019774
+:10028000F1F700C000008FE0809308018091060123
+:100290008F708F3051F30895D1DF89EF90E001978F
+:1002A000F1F700C00000E1DFADDF90E0FC01E757AF
+:1002B000F109E837F10518F58827EE58FF4F8F4F01
+:1002C000A7C18AE2089584E3089581E3089580E355
+:1002D000089588E3089585E3089582E3089583E20D
+:1002E000089589E3089586E3089583E3089584E4F7
+:1002F000089583E4089582E4089581E408958AE7E7
+:10030000089587E3089585B1806285B9000085B1BD
+:100310008F7D85B90895CF93C82F662331F0E2E031
+:10032000F1E080818064808305C0E2E0F1E08081BB
+:100330008F7B808385B18F7E85B98DB180628DB9C9
+:1003400083B3806283BB8DB188608DB9E1E0F1E059
+:10035000808188608083E2E0F1E08081877F808314
+:100360008EB1877F8EB984B38F7D84BB8EB18F7D34
+:100370008EB9CC231CF4808188608083C6FF03C0C3
+:100380008EB188608EB9C5FF03C084B3806284BB20
+:10039000C4FF03C08EB180628EB9B5DFE2E0F1E048
+:1003A0008081877F80838EB1877F8EB984B38F7D74
+:1003B00084BB8EB18F7D8EB9C3FF03C080818860FE
+:1003C0008083C2FF03C08EB188608EB9C1FF03C0B5
+:1003D00084B3806284BBC0FF03C08EB180628EB9DB
+:1003E00092DF8EB180628EB984B3806284BB8EB19D
+:1003F00088608EB9E2E0F1E0808188608083CF91EF
+:100400000895882331F0E2E0F1E080818064808308
+:1004100005C0E2E0F1E080818F7B808385B180615F
+:1004200085B98DB18F7D8DB983B38F7D83BB8DB140
+:10043000877F8DB9E1E0F1E08081877F808385B19E
+:10044000806285B90000659902C080E001C080E14A
+:1004500095998062639980649091000193FD806812
+:1004600095B19F7D95B9000095B1906295B9000056
+:10047000659981609599826063998460909100018B
+:1004800093FD886095B19F7D95B9089580E0B9DFAF
+:100490008823E4F300C000C080E0B3CF0895CF9379
+:1004A000C82FF4DF60E08C2F36DFCF910895611103
+:1004B00003C08058F4CF08958054F1CF089581E0AF
+:1004C000EECF0895CF93C82FE1DFCA3041F48034D6
+:1004D00010F480E401C080E08058E1DF03C061E0F7
+:1004E0008C2F19DFCF910895CF93DF93FC0102C0C9
+:1004F000E9DFFE01EF01219680818111F9CFDF91C3
+:10050000CF910895CF93C82FE1E0F1E0808180641E
+:10051000808384B1806184B984B1806284B98DB1F3
+:1005200080628DB983B3806283BB8DB188608DB9E1
+:100530008081886080838FE99FE00197F1F700C098
+:10054000000084B3806284BB8EB180628EB9DBDE32
+:1005500081EE94E00197F1F700C00000D4DE95E150
+:100560009A95F1F70000CFDE85E18A95F1F700005A
+:100570008EB18F7D8EB9C7DE95E19A95F1F70000B7
+:1005800088E28DDF88E08BDF9ADF86E088DF8C2FC2
+:1005900086DFCF9108958CE0B5DF91DF80E092E0B7
+:1005A000A3DF2FEF83ED90E3215080409040E1F7EF
+:1005B00000C000003FCE0895CF93DF9300D01F927C
+:1005C000CDB7DEB7E8DF68DE2FEF35EA9EE02150D9
+:1005D00030409040E1F700C000008F3FA1F39FEC56
+:1005E000980F993040F480534AE0BE016F5F7F4F0F
+:1005F00090E016D002C089831A8261DF86E092E023
+:1006000073DF61E080E053DFCE0101966DDFDBCF69
+:10061000EE0FFF1F881F8BBF0790F691E02D1994F6
+:10062000BB274A3031F4992322F4BDE2909581959D
+:100630009F4F01C0BB27FB015527AA27880F991F91
+:10064000AA1FA41710F0A41B83955051B9F7A05D01
+:10065000AA3308F0A95DA193009779F7B111B1937E
+:100660001192CB0100C0DC01FC01672F719177234F
+:10067000E1F7329704C07C916D9370836291AE175D
+:0A068000BF07C8F30895F894FFCFF8
+:10068A005265616479004B657920507265737365B0
+:04069A00643A0000BE
+:00000001FF
