+++ TODO.md
@@ -1,0 +1,20 @@
+### implemented so far:
+
+---
+
+- Wiring and hardware
+
+---
+
+- MEGA keypad, screen
+- MEGA SPI ISR send
+
+---
+
+- uno SPI ISR Recieve
+
+### todo:
+
+- uno libraries
+- uno leds and buzzer control.
+- Sending of correct data with SPI.
