--- uno/unolib.c
@@ -1,22 +1,0 @@
-#include "uno_lib.h"
-#include <avr/io.h>
-#include <util/delay.h>
-
-void uno_init(void) {
-    // Example: Configure PD2 for the movement LED, PD3 for the door LED, PD4 for the buzzer.
-    // Set PD2, PD3, and PD4 as output pins.
-    DDRD |= (1 << PD2) | (1 << PD3) | (1 << PD4);
-    
-    // Initialize outputs to off
-    PORTD &= ~((1 << PD2) | (1 << PD3) | (1 << PD4));
-}
-
-void uno_execute_movement(void) {
-    // Example: Blink the movement LED to simulate elevator movement.
-    PORTD |= (1 << PD2);  // Turn ON movement LED.
-    _delay_ms(500);
-    PORTD &= ~(1 << PD2); // Turn OFF movement LED.
-    _delay_ms(500);
-    
-    // Add additional control logic for door LED and buzzer as needed.
-}
